retrain-job/requirements.txt
catboost==1.2
pandas==2.1.4
psycopg2-binary==2.9.9
Requests==2.31.0
